feat: add class for calculating the location of TaskScheduler/ScheduledExecutor beans
Scheduled Tasks can be configured with a TaskScheduler directly on the ScheduledTaskRegistrar, with
a custom TaskScheduler bean, or even with a ScheduledExecutorService bean. As bugsnag sets an
ErrorHandler on the TaskScheduler to capture unhandled exceptions in Scheduled Tasks, we need to
calculate the location of this bean as part of configuration. The approach used broadly follows
Spring's implementation - we search the default BeanFactory for beans of the given type, and use the
'taskScheduler' bean name as an additional filter if more than one bean is found.