fix: respect different methods of configuring a TaskScheduler
Alters the ScheduledTaskConfiguration so that it takes into account any TaskScheduler or
ScheduledExecutorService beans, rather than ovverriding these with a different TaskScheduler. The
configuration will now use any scheduler on the TaskRegistrar, then any TaskScheduler bean, then any
ScheduledExecutorService bean, and finally will create its own TaskScheduler. In all cases we
attempt to set an ErrorHandler on the TaskScheduler so that any unhandled exceptions in scheduled
tasks are reported.