Allow a BugsnagAppender to be created from an existing client.
This makes it possible to programatically create a Bugsnag logback appender without having to pass
in the configuration options twice (once to the original client, and once again to the appender).
This is especially useful if you are adding callbacks and configuration settings that are the same
across the original instance and the Logback appender.

If the appender is created with an existing client, it is not possible to set any of the values of
the appender that affect Bugsnag as it would also affect the original client.