fix: jvm hangs when System.exit or bugsnag.close is not called
The fix to prevent a resource leak in #143 removed the task scheduler running as
a daemon thread and so doesn't exit unless the application is explicitly exited
(via System.exit) or the thread is explicitly stopped (via bugsnag.close).

This fix should be the benefits of both #143 and #121 to gracefully shutdown in
both use cases.

Updated the Spring Boot app test and example to remove explict System.exit call
as this should not be required.

fixes #151