test(mazerunner): fix failing tests for plain spring app due to tomcat root change
It appears that Tomcat 7.0 no longer has a ROOT webapp and so serves a 404 instead. This was being
used by Mazerunner to check whether the app had started yet. However I believe it's more correct
to check the deployed application rather than this placeholder app.

Therefore this changeset deploys the spring app to the root and amends the tests to run from the
root.

A consequence of this is that the automatic session tracking tests are affected as each test starts
by accessing a page and incrementing the counts. Therefore the scenario's expectation has been
updated (and annotated) to expect a second count.

fix